+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The **patch** part changes incrementally at each release.
 * mysql: install save_mysql_processlist script
 * nextcloud: New role to setup a nextcloud instance
 * redis: variable to force use of port 6379 in instances mode
+* redis: check maxmemory in NRPE check
 * lxc-php: Allow php containers to contact local MySQL with localhost
 
 ### Changed